Develop

+++ src/main/java/com/jaksim3/bak/service/CartService.java
@@ -28,7 +28,7 @@ public class CartService {
     private final CartProductRepository cartProductRepository;
 
     @Transactional
-    public ProductDto.Response save(Long productId) {
+    public ProductDto.ProductResponse save(Long productId) {
         Member member = memberRepository.findById(SecurityUtil.getCurrentMemberId()).orElseThrow(
                 () -> new IllegalArgumentException("해당 회원이 없습니다"));
         Product product = productRepository.findById(productId).orElseThrow(
@@ -48,28 +48,30 @@ public class CartService {
         product.addCartProduct(cartProduct);
         cartProductRepository.save(cartProduct);
 
-        return ProductDto.Response.of(product);
+        return ProductDto.ProductResponse.of(product);
     }
 
     @Transactional
-    public Long delete(Long productId) {
+    public void delete(Long productId) {
+        Member member = memberRepository.findById(SecurityUtil.getCurrentMemberId()).orElseThrow(
+                () -> new IllegalArgumentException("해당 회원이 없습니다"));
         Product product = productRepository.findById(productId).orElseThrow(
                 () -> new IllegalArgumentException("해당 상품이 없습니다"));
-        CartProduct cartProduct = cartProductRepository.findByProduct(product).orElseThrow(
+        CartProduct cartProduct = cartProductRepository.findByProductAndCart(product,member.getCart()).orElseThrow(
                 () -> new IllegalArgumentException("해당 장바구니 상품이 없습니다"));
 
-        cartProductRepository.delete(cartProduct);
-
-        return cartProduct.getProduct().getId();
+        if(member.getCart().getCartProducts().contains(cartProduct)) {
+            cartProductRepository.delete(cartProduct);
+        }
     }
 
-    public List<ProductDto.Response> getCartProductList() {
+    public List<ProductDto.ProductResponse> getCartProductList() {
         Member member = memberRepository.findById(SecurityUtil.getCurrentMemberId()).orElseThrow(
                 () -> new IllegalArgumentException("해당 회원이 없습니다"));
 
         return cartProductRepository.findAllByCart(member.getCart())
                 .stream()
-                .map(cartProduct -> ProductDto.Response.of(cartProduct.getProduct()))
+                .map(cartProduct -> ProductDto.ProductResponse.of(cartProduct.getProduct()))
                 .collect(Collectors.toList());
     }
